// components/SmartSearch/SmartSearch.module.css.d.ts
declare const styles: {
  searchContainer: string;
  searchBar: string;
  focused: string;
  searchIcon: string;
  searchInput: string;
  clearButton: string;
  filterButton: string;
  activeFilter: string;
  searchButton: string;
  suggestionsContainer: string;
  suggestionItem: string;
  suggestionHeader: string;
  filtersContainer: string;
  filterSection: string;
  filterOptions: string;
  filterOption: string;
  ratingFilter: string;
  starButton: string;
  tagCloud: string;
  tag: string;
  activeTag: string;
};

export default styles;